use tsup for building, migrate to vite, add default handlers
