+++ packages/docs/src/demos/react-query/App.test.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import { composeStories } from '@storybook/react'
+import { vi } from 'vitest'
 
 import { getServer } from '../../test-utils'
 import * as stories from './App.stories'
@@ -10,7 +11,7 @@ const { MockedSuccess, MockedError } = composeStories(stories)
 const server = getServer()
 
 afterAll(() => {
-  jest.restoreAllMocks()
+  vi.restoreAllMocks()
 })
 
 it('renders film cards for each film', async () => {
@@ -32,7 +33,7 @@ it('renders film cards for each film', async () => {
 
 it('renders error message if it cannot load the films', async () => {
   // get rid of the console.error for this test which adds clutter to the logs
-  jest.spyOn(console, 'error').mockImplementationOnce(() => {})
+  vi.spyOn(console, 'error').mockImplementationOnce(() => {})
 
   server.use(...MockedError.parameters.msw.handlers)
   render(<MockedError />)
