+++ packages/docs/vite.config.js
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    clearMocks: true,
+    setupFiles: './src/setupTests.js',
+  },
+})
