+++ packages/msw-addon/tsup.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig((options) => ({
+  entry: ['src/index.ts'],
+  splitting: false,
+  minify: false,
+  format: ['cjs', 'esm'],
+  dts: {
+    resolve: true,
+  },
+  treeshake: true,
+  clean: true,
+  platform: 'browser',
+  esbuildOptions(options) {
+    options.conditions = ['module']
+  },
+}))
